обновляет е2е тесты для маленькой карточки фильма
